+++ ember-cli-build.js
@@ -14,13 +14,15 @@ module.exports = function (defaults) {
     behave. You most likely want to be modifying `./index.js` or app's build file
   */
 
-  const { maybeEmbroider } = require('@embroider/test-setup');
-  return maybeEmbroider(app, {
-    skipBabel: [
-      {
-        package: 'qunit',
-      },
-      { package: 'velocity' },
-    ],
-  });
+  return app.toTree();
+
+  // const { maybeEmbroider } = require('@embroider/test-setup');
+  // return maybeEmbroider(app, {
+  //   skipBabel: [
+  //     {
+  //       package: 'qunit',
+  //     },
+  //     { package: 'velocity' },
+  //   ],
+  // });
 };